Move utils methods/types to separate files
